Move dag status type declaration
